+++ STACpopulator/implementations/CORDEXCMIP6_Ouranos/add_CORDEX6.py
@@ -1,18 +1,17 @@
-import argparse
 import logging
 from typing import Any
 
-from requests.sessions import Session
-
 from STACpopulator.extensions.cordex6 import Cordex6DataModel
-from STACpopulator.input import ErrorLoader, THREDDSLoader
-from STACpopulator.populator_base import STACpopulatorBase
+from STACpopulator.populators import THREDDSPopulator
 
 LOGGER = logging.getLogger(__name__)
 
 
-class CORDEX_STAC_Populator(STACpopulatorBase):
+class CORDEX_STAC_Populator(THREDDSPopulator):
     """Populator that creates STAC objects representing CORDEX data from a THREDDS catalog."""
+
+    name = "CORDEXCMIP6_Ouranos"
+    description = "CMIP6-CORDEX STAC populator from a THREDDS catalog or NCML XML."
 
     data_model = Cordex6DataModel
 
@@ -20,49 +19,3 @@ class CORDEX_STAC_Populator(STACpopulatorBase):
         """Return a STAC item."""
         dm = self.data_model.from_data(item_data)
         return dm.stac_item()
-
-
-# TODO: This probably doesn't need to be copied for every implementation, right ?
-def add_parser_args(parser: argparse.ArgumentParser) -> None:
-    """Add additional CLI arguments to the argument parser."""
-    parser.description = "CMIP6-CORDEX STAC populator from a THREDDS catalog or NCML XML."
-    parser.add_argument("stac_host", help="STAC API URL")
-    parser.add_argument("href", help="URL to a THREDDS catalog or a NCML XML with CMIP6 metadata.")
-    parser.add_argument("--update", action="store_true", help="Update collection and its items")
-    parser.add_argument(
-        "--mode",
-        choices=["full", "single"],
-        default="full",
-        help="Operation mode, processing the full dataset or only the single reference.",
-    )
-    parser.add_argument(
-        "--config",
-        type=str,
-        help=(
-            "Override configuration file for the populator. "
-            "By default, uses the adjacent configuration to the implementation class."
-        ),
-    )
-
-
-def runner(ns: argparse.Namespace, session: Session) -> int:
-    """Run the populator."""
-    LOGGER.info(f"Arguments to call: {vars(ns)}")
-
-    if ns.mode == "full":
-        data_loader = THREDDSLoader(ns.href, session=session)
-    else:
-        # To be implemented
-        data_loader = ErrorLoader()
-
-    c = CORDEX_STAC_Populator(
-        ns.stac_host,
-        data_loader,
-        update=ns.update,
-        session=session,
-        config_file=ns.config,
-        update_collection=ns.update_collection,
-        exclude_summaries=ns.exclude_summary,
-    )
-    c.ingest()
-    return 0