+++ STACpopulator/implementations/DirectoryLoader/__init__.py
@@ -1,3 +1,0 @@
-from STACpopulator.implementations.DirectoryLoader.crawl_directory import add_parser_args, runner
-
-__all__ = ["add_parser_args", "runner"]