+++ STACpopulator/populators.py
@@ -1,3 +1,6 @@
+from __future__ import annotations
+
+import argparse
 import functools
 import importlib
 import importlib.util
@@ -8,9 +11,10 @@ import os
 import re
 from abc import ABC, abstractmethod
 from datetime import datetime
-from typing import Any, Callable, Dict, Iterable, List, MutableMapping, Optional, Union
+from typing import Any, Callable, Dict, Iterable, List, MutableMapping, Optional, Union, get_args
 
 import pystac
+import requests
 from requests.sessions import Session
 
 from STACpopulator.api_requests import (
@@ -21,7 +25,7 @@ from STACpopulator.api_requests import (
 )
 from STACpopulator.collection_update import UpdateModesOptional, update_collection
 from STACpopulator.exceptions import FunctionLoadError
-from STACpopulator.input import GenericLoader
+from STACpopulator.input import ErrorLoader, GenericLoader, THREDDSLoader
 from STACpopulator.stac_utils import load_config
 
 LOGGER = logging.getLogger(__name__)
@@ -29,6 +33,9 @@ LOGGER = logging.getLogger(__name__)
 
 class STACpopulatorBase(ABC):
     """Abstract base class for STAC populators."""
+
+    name: Optional[str]
+    description: Optional[str]
 
     def __init__(
         self,
@@ -78,6 +85,13 @@ class STACpopulatorBase(ABC):
         LOGGER.info("Initialization complete")
         LOGGER.info(f"Collection {self.collection_name} is assigned ID {self.collection_id}")
         self._collection = self.create_stac_collection()
+
+    @classmethod
+    def concrete_subclasses(cls) -> list[STACpopulatorBase]:
+        """Return all concrete subclasses (recursive) of this class."""
+        return [
+            k for klass in cls.__subclasses__() for k in [klass] + klass.__subclasses__() if not inspect.isabstract(k)
+        ]
 
     @staticmethod
     def _load_extra_parser(func_str: str, extra_kwargs: dict[str, str]) -> Callable:
@@ -314,3 +328,102 @@ class STACpopulatorBase(ABC):
             LOGGER.info(f"Processed {counter} data items. {failures} failures")
         if self.update and self.update_collection != "none":
             self.publish_stac_collection(self._collection)
+
+    @staticmethod
+    def _extra_parser_argument(arg: str) -> tuple[str, str]:
+        if "=" in arg:
+            return tuple(a.strip() for a in arg.split("=", 1))
+        raise argparse.ArgumentTypeError("--extra-parser-arguments must be in the form 'key=value'")
+
+    @classmethod
+    def update_parser_args(cls, parser: argparse.ArgumentParser) -> None:
+        """Add additional CLI arguments to the argument parser."""
+        parser.add_argument("stac_host", help="STAC API URL")
+        parser.add_argument("--update", action="store_true", help="Update collection and its items")
+        parser.add_argument(
+            "--update-collection-mode",
+            dest="update_collection",
+            choices=get_args(UpdateModesOptional),
+            default="none",
+            help="Update collection information based on new items created or updated by this populator. "
+            "Only applies if --update is also set.",
+        )
+        parser.add_argument(
+            "--exclude-summary",
+            nargs="*",
+            action="extend",
+            default=[],
+            help="Exclude these properties when updating collection summaries. ",
+        )
+        parser.add_argument(
+            "-x",
+            "--extra-item-parsers",
+            action="append",
+            help="Functions that may modify items before upload. "
+            "Should be specified in the form 'module:function_name' "
+            "and have the signature function(item: dict, **kw)",
+        )
+        parser.add_argument(
+            "-X",
+            "--extra-collection-parsers",
+            action="append",
+            help="Functions that may modify collections before upload. "
+            "Should be specified in the form 'module:function_name' or "
+            "path/to/python/file.py:function_name. Functions should "
+            "have the signature function(collection: dict, **kw) -> None "
+            "and should modify the collection dict in place.",
+        )
+        parser.add_argument(
+            "-a",
+            "--extra-parser-arguments",
+            action="append",
+            type=cls._extra_parser_argument,
+            help="Extra keyword arguments that should be passed to extra "
+            "item and collection function as "
+            "keyword arguments. "
+            "Should be specified in the form 'key=value'",
+        )
+
+    @classmethod
+    @abstractmethod
+    def run(cls, ns: argparse.Namespace, session: requests.Session) -> int:
+        """Run the populator given the arguments from the command line."""
+        raise NotImplementedError
+
+
+class THREDDSPopulator(STACpopulatorBase):
+    """Base class for populators that get data from a THREDDS catalog."""
+
+    @classmethod
+    def update_parser_args(cls, parser: argparse.ArgumentParser) -> None:
+        """Add additional CLI arguments to the argument parser."""
+        super().update_parser_args(parser)
+        parser.add_argument("href", help="URL to a THREDDS catalog or a NCML XML with CMIP6 metadata.")
+        parser.add_argument(
+            "--mode",
+            choices=["full", "single"],
+            default="full",
+            help="Operation mode, processing the full dataset or only the single reference.",
+        )
+        parser.add_argument(
+            "--config",
+            type=str,
+            help=(
+                "Override configuration file for the populator. "
+                "By default, uses the adjacent configuration to the implementation class."
+            ),
+        )
+
+    @classmethod
+    def run(cls, ns: argparse.Namespace, session: requests.Session) -> int:
+        """Run the populator given the arguments from the command line."""
+        LOGGER.info(f"Arguments to call: {vars(ns)}")
+
+        if ns.mode == "full":
+            data_loader = THREDDSLoader(ns.href, session=session)
+        else:
+            # To be implemented
+            data_loader = ErrorLoader()
+
+        cls(ns.stac_host, data_loader, update=ns.update, session=session, config_file=ns.config).ingest()
+        return 0